+++ lib/ui/academics/AcademicsLinks.dart
@@ -1,6 +1,7 @@
 
 import 'package:illinois/model/Analytics.dart';
 import 'package:illinois/ui/gbv/GBVResourceDirectoryContentPanel.dart';
+import 'package:illinois/ui/home/HomeSavedResourcesWidget.dart';
 import 'package:rokwire_plugin/service/localization.dart';
 
 class AcademicLinksPanel extends GBVResourceDirectoryContentPanel {
@@ -15,6 +16,8 @@ class AcademicLinksWidget extends GBVResourceDirectoryContentWidget {
   AcademicLinksWidget() : super(
     contentCategory: 'academic_links',
     contentAssetKey: 'assets/extra/academicLinks.json',
+    favoriteKey: HomeSavedResourcesWidget.favoriteKey,
+    favoriteListner: HomeSavedResourcesWidget.favoriteListener,
     contentFailedMessage: Localization().getStringEx('', 'Failed to load academic links data'),
   );
 }