+++ lib/ui/home/HomeFavoritesWidget.dart
@@ -74,7 +74,7 @@ class HomeFavoritesWidget extends StatefulWidget {
       case LaundryRoom.favoriteKeyName: return Localization().getStringEx('widget.home.favorites.title.laundry', 'My Laundry');
       case MTDStop.favoriteKeyName: return Localization().getStringEx('widget.home.favorites.title.mtd_stops', 'My Bus Stops');
       case ExplorePOI.favoriteKeyName: return Localization().getStringEx('widget.home.favorites.title.my_locations', 'My Locations');
-      case GuideFavorite.favoriteKeyName: return Localization().getStringEx('widget.home.favorites.title.campus_guide', 'Saved Resources');
+      case GuideFavorite.favoriteKeyName: return Localization().getStringEx('widget.home.favorites.title.campus_guide', 'My University Living');
       case Appointment.favoriteKeyName: return Localization().getStringEx('widget.home.favorites.title.appointments', 'MyMcKinley Appointments');
     }
     return null;
@@ -428,30 +428,18 @@ class _HomeFavoritesWidgetState extends State<HomeFavoritesWidget> with Notifica
     CollectionUtils.isNotEmpty(favoriteIds) ? ListUtils.reversed(ExplorePOI.listFromString(favoriteIds)) : null;
 
   Future<List<Favorite>?> _loadFavoriteGuideItems(LinkedHashSet<String>? favoriteIds) async {
-    List<Favorite>? guideItems;
     if ((favoriteIds != null) && (Guide().contentList != null)) {
-      
-      Map<String, Favorite> favorites = <String, Favorite>{};
-      for (dynamic contentEntry in Guide().contentList!) {
-        String? guideEntryId = Guide().entryId(JsonUtils.mapValue(contentEntry));
-        
-        if ((guideEntryId != null) && favoriteIds.contains(guideEntryId)) {
-          favorites[guideEntryId] = GuideFavorite(id: guideEntryId);
+      List<Favorite> result = <Favorite>[];
+      for (String favoriteId in favoriteIds) {
+        GuideFavorite guideEntryFavorite = GuideFavorite.fromString(favoriteId);
+        if (Guide().entryById(guideEntryFavorite.id) != null) {
+          result.add(guideEntryFavorite);
         }
       }
-
-      if (favorites.isNotEmpty) {
-        List<Favorite> result = <Favorite>[];
-        for (String favoriteId in favoriteIds) {
-          Favorite? favorite = favorites[favoriteId];
-          if (favorite != null) {
-            result.add(favorite);
-          }
-        }
-        guideItems = List.from(result.reversed);
-      }
+      return result.reversed.toList();
+    } else {
+      return null;
     }
-    return guideItems;
   }
 
   Future<List<Favorite>?> _loadFavoriteAppointments(LinkedHashSet<String>? favoriteIds) async =>
@@ -514,7 +502,7 @@ class _HomeFavoritesWidgetState extends State<HomeFavoritesWidget> with Notifica
       case LaundryRoom.favoriteKeyName: return Localization().getStringEx('widget.home.favorites.message.offline.laundry', 'My Laundry are not available while offline.');
       case MTDStop.favoriteKeyName: return Localization().getStringEx('widget.home.favorites.message.offline.mtd_stops', 'My Bus Stops are not available while offline.');
       case ExplorePOI.favoriteKeyName: return Localization().getStringEx('widget.home.favorites.message.offline.locations', 'My Locations are not available while offline.');
-      case GuideFavorite.favoriteKeyName: return Localization().getStringEx('widget.home.favorites.message.offline.campus_guide', 'Saved Resources are not available while offline.');
+      case GuideFavorite.favoriteKeyName: return Localization().getStringEx('widget.home.favorites.message.offline.campus_guide', 'My University Living are not available while offline.');
       case Appointment.favoriteKeyName: return Localization().getStringEx('widget.home.favorites.message.offline.appointments', 'MyMcKinley Appointments are not available while offline.');
     }
     return null;